[TASK] Testing PageRenderer integration, switching to page. TS cObj
