+++ Classes/Core/ViewHelper/AbstractPageRendererViewHelper.php
@@ -25,12 +25,15 @@ namespace TYPO3\Beautyofcode\Core\ViewHelper;
  * This copyright notice MUST APPEAR in all copies of the script!
  ***************************************************************/
 
+use TYPO3\CMS\Frontend\Controller\TypoScriptFrontendController;
+
 /**
  * Abstract page renderer based view helper
  *
  * @package \TYPO3\Beautyofcode\Core\ViewHelper
  * @author Thomas Juhnke <typo3@van-tomas.de>
- * @license http://www.gnu.org/licenses/gpl.html GNU General Public License, version 3 or later
+ * @license http://www.gnu.org/licenses/gpl.html
+ *          GNU General Public License, version 3 or later
  * @link http://www.van-tomas.de/
  */
 abstract class AbstractPageRendererViewHelper extends \TYPO3\CMS\Fluid\Core\ViewHelper\AbstractViewHelper {
@@ -43,10 +46,27 @@ abstract class AbstractPageRendererViewHelper extends \TYPO3\CMS\Fluid\Core\View
 
 	/**
 	 *
+	 * @var TypoScriptFrontendController
+	 */
+	protected $fe;
+
+	/**
+	 * injectPageRenderer
+	 *
 	 * @param \TYPO3\CMS\Core\Page\PageRenderer $pageRenderer
+	 * @return void
 	 */
 	public function injectPageRenderer(\TYPO3\CMS\Core\Page\PageRenderer $pageRenderer) {
 		$this->pageRenderer = $pageRenderer;
+	}
+
+	/**
+	 * initializeObject
+	 *
+	 * @return void
+	 */
+	public function initializeObject() {
+		$this->fe = $GLOBALS['TSFE'];
 	}
 }
 ?>
