+++ Configuration/TypoScript/Prism/setup.txt
@@ -29,3 +29,5 @@ plugin.tx_beautyofcode {
 		.tx_beautyofcode pre { overflow: auto; }
 	)
 }
+
+page.1407710024 < tt_content.list.20.beautyofcode_pageassets
