--- Resources/Private/BrushLoader/Layouts/Default.html
@@ -1,3 +1,0 @@
-<f:layout name="Default" />
-
-<f:render section="Main" />
