+++ Resources/Private/Layouts/PageAssets.html
@@ -1,0 +1,3 @@
+<f:layout name="PageAssets" />
+
+<f:render section="Main" />
