+++ Resources/Private/Partials/PageAssets/Prism.html
@@ -1,5 +1,4 @@
 {namespace boc=TYPO3\Beautyofcode\ViewHelpers}
-<f:layout name="Default" />
 
 <f:section name="Main">
 	<f:for each="{brushes}" as="brush">
