+++ Resources/Private/Partials/PageAssets/SyntaxHighlighter.html
@@ -1,5 +1,4 @@
 {namespace boc=TYPO3\Beautyofcode\ViewHelpers}
-<f:layout name="Default" />
 
 <f:section name="Main">
 	<boc:pageRenderer.addJsFooterInlineCode name="boc_inline">
