+++ Resources/Private/Templates/PageAssets/Add.html
@@ -1,0 +1,5 @@
+<f:layout name="PageAssets" />
+
+<f:section name="Main">
+	<f:render partial="PageAssets/{settings.library}" section="Main" arguments="{_all}" />
+</f:section>
